pyproject.toml
[tool.poetry]
name = "gen-ai"
version = "0.1.0"
description = ""
authors = ["hassanrzprop <hassanrzprop@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
langchain-huggingface = "^0.1.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.10"
langchain-google-genai = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
